[common/prost-ext] rename to prost-test-helpers
This is a test-only crate, so ensure that it is reflected in the name.
