+++ grpc/types/src/unit_tests/ledger_info_test.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use libra_prost_ext::test_helpers::assert_protobuf_encode_decode;
+use libra_prost_test_helpers::assert_protobuf_encode_decode;
 use libra_types::ledger_info::{LedgerInfo, LedgerInfoWithSignatures};
 use proptest::prelude::*;
