+++ types/Cargo.toml
@@ -39,7 +39,7 @@ regex = "1.3.7"
 proptest = "0.9.6"
 proptest-derive = "0.1.2"
 libra-proptest-helpers = { path = "../common/proptest-helpers", version = "0.1.0" }
-libra-prost-ext = { path = "../common/prost-ext", version = "0.1.0" }
+libra-prost-test-helpers = { path = "../common/prost-test-helpers", version = "0.1.0" }
 serde_json = "1.0.53"
 
 [features]
